move machine slice constants to separate file
